+++ lib/dicey.rb
@@ -19,7 +19,6 @@ module Dicey
   require_relative "dicey/die_foundry"
 
   require_relative "dicey/distribution_properties_calculator"
-  require_relative "dicey/roller"
   Dir["dicey/sum_frequency_calculators/*.rb", base: __dir__].each { require_relative _1 }
 
   require_relative "dicey/version"